+++ R/langevitour.R
@@ -4,7 +4,9 @@
 #'
 #' To retain the original units on plot axes within the widget, use \code{center} and \code{scale} rather than altering X.
 #'
-#' langevitour will by default not scale variables individually. If you want this, use something like \code{scale=apply(X,2,sd)*4}.
+#' langevitour will by default not scale variables individually. If you want variables to be individually scaled, use something like \code{scale=apply(X,2,sd)*4}.
+#'
+#' In Javascript, the langevitour object can be obtained using \code{document.getElementById(elementId).langevitour}. For example you could have a button that sets the state of a widget using \code{document.getElementById(elementId).langevitour.setState(desiredState)}.
 #'
 #' @param X The data to plot. A matrix of numeric data, or something that can be cast to a matrix. Rows will be shown as points in the widget. Columns are the variables of your data.
 #'
@@ -18,9 +20,21 @@
 #'
 #' @param extraAxes A matrix with each column defining a projection of interest. The columns of \code{X \%*\% extraAxes} will be presented as extra "variables".
 #'
+#' @param lineFrom A vector of row numbers. Draw lines starting at these rows.
+#'
+#' @param lineTo A vector of row numbers. Draw lines ending at these rows.
+#'
 #' @param axisColors Character vector. Colors for each variable and then each extra axis.
 #'
+#' @param levelColors Character vector. Colors for each level of \code{group}.
+#'
+#' @param colorVariation Number between 0 and 1. Individual points are given slightly different brightnesses. How strong should this effect be?
+#'
 #' @param pointSize Point radius in pixels.
+#'
+#' @param subsample For speed, randomly subsample down to this many rows.
+#'
+#' @param state A JSON string, or an object that htmlwidgets will convert to the correct JSON. Initial widget state settings. The state of a widget can be obtained by pressing the "?" button. I am not going to guarantee that states will be compatible between versions of langevitour.
 #'
 #' @param width Width of widget.
 #'
@@ -46,8 +60,9 @@
 #' @export
 langevitour <- function(
         X, group=NULL, name=NULL, center=NULL, scale=NULL, 
-        extraAxes=NULL, axisColors=NULL, pointSize=1,
-        width=NULL, height=NULL, elementId=NULL) {
+        extraAxes=NULL, lineFrom=NULL, lineTo=NULL,
+        axisColors=NULL, levelColors=NULL, colorVariation=0.3, pointSize=1, subsample=NULL, 
+        state=NULL, width=NULL, height=NULL, elementId=NULL) {
 
     X <- as.matrix(X)
     
@@ -97,6 +112,26 @@ langevitour <- function(
     names(scale) <- NULL
     names(group) <- NULL
     
+    # Subsampling
+    
+    if (!is.null(subsample) && subsample < nrow(X)) {
+        ind <- sample.int(nrow(X), subsample)
+        
+        X_centered_scaled <- X_centered_scaled[ind,,drop=FALSE]
+        group <- group[ind]
+        if (!is.null(name))
+            name <- name[ind]
+        if (!is.null(lineFrom)) {
+            lineFrom <- match(lineFrom, ind)
+            lineTo <- match(lineTo, ind)
+            keep <- !is.na(lineFrom) & !is.na(lineTo)
+            lineFrom <- lineFrom[keep]
+            lineTo <- lineTo[keep]
+        }
+    }
+    
+    # Convert to form that will JSON-ify correctly
+    
     data <- list(
         X = X_centered_scaled,
         center = as.list(as.numeric(center)),
@@ -110,8 +145,16 @@ langevitour <- function(
         extraAxesCenter = as.list(as.numeric(extraAxesCenter)),
         extraAxesNames = as.list(colnames(extraAxes)),
         
+        # Convert from 1 based to 0 based indices
+        lineFrom = as.list(as.numeric(lineFrom) - 1), 
+        lineTo = as.list(as.numeric(lineTo) - 1),
+        
         axisColors=as.list(as.character(axisColors)),
-        pointSize=as.numeric(pointSize))
+        levelColors=as.list(as.character(levelColors)),
+        colorVariation=as.numeric(colorVariation),
+        pointSize=as.numeric(pointSize),
+        
+        state=state)
     
     htmlwidgets::createWidget(
         name = 'langevitour',