+++ NEWS.md
@@ -1,3 +1,16 @@
+# langevitour 0.4
+
+* Fix README.md, remove redundant brackets.
+
+# langevitour 0.3
+
+* Ability to get and set state with Javascript.
+* Can now draw lines between points.
+* Axis deactivation is now animated.
+* Added ultralocal point repulsion.
+* Level colors can be specified.
+* Add knnDenoise function.
+
 # langevitour 0.2
 
 * Remove RNA-Seq example from CRAN package -- dependency not available on CRAN build servers.